requirements.txt
numpy
flask
graphviz
pyclipper
flask_socketio
scipy
sklearn
python-socketio<4
eventlet
